feat(sequenceEqual): add higher-order lettable version of sequenceEqual
